Split parseCLI into header and definition
